feat: Add job run stats reporting (#517)
* fix: Update download files to return JSON errors

Our implementation did not integrate with the GetHandler helper
presumably because the success case is streaming a file. The update
returns an empty object on success, which does not matter to the job
creator but helps us integrate with GetHandler.

* feat: Add stats options

* feat: Add stats to solver server

* feat: Add HTTPStats client options

* feat: Add GetShortcut helper to get module identifier

* feat: Add job run stats

We post job run details to the stats API, including job run time. The
job run time starts when the job offer is created and ends when the job
creator finishes downloading the job outputs.

* feat: Add reputation stats

* feat: Post job completed no validation

* chore: Update job offer CreatedAt comment

* chore: Debug log download file signature-address mismatches

* chore: Warn log when an expected file is missing in download files handlers